Move recipe.proto to java/arcs/core/data/proto/
This would make it easier to organize utilities around the proto.
